Use multistage build to leave out libpostal buildtime dependencies
With https://github.com/pelias/docker-baseimage/pull/23 removing the C++
compiler toolchain from our baseimages, ideally we will remove build
time dependencies everywhere to save space on disk and over the network.

The libpostal baseimage does require a compiler toolchain to build
libpostal, but not to run it. So we can use a multistage image where
libpostal is compiled with all its dependencies, but only the build
artefacts are kept in the final image.

That saves a few hundred MB, but the libpostal GitHub repository is also
about 80MB, so we get even more savings there.