Fix usage of `DESTDIR` and `--prefix`
One last (hopefully) fix to get everything working. (My Makefile skills
are a little rusty).

Instead of using the `--prefix` option to `./configure` (which tells
`./configure) the final location that the package will actually be
installed into, use `PREFIX` when calling make install. This basically
translates all the directories without changing anything else.

This is what we want, since the idea is to create a directory with just
libpostal files that can be cleanly copied from the builder image to the
final image.